requirements.txt
# This file is autogenerated. Do not edit it manually.
# Run `make freeze-requirements` to update requirements.txt
# with package version changes made in requirements-app.txt

Flask==1.0.3
Flask-Env==2.0.0

boto3==1.9.64

python-dotenv==0.10.3
python-magic==0.4.15

# PaaS

gunicorn==19.7.1  # pyup: ignore # >19.8 stops eventlet workers after a timeout
eventlet==0.24.1

awscli-cwlogs>=1.4,<1.5

gds-metrics==0.2.0

git+https://github.com/cds-snc/notifier-utils.git@34.0.11#egg=notifications-utils==34.0.11

socketio-client==0.5.6
requests
requests[security]
pycryptodome

git+https://bitbucket.org/cse-assemblyline/assemblyline_client.git@v3.7.3#egg=assemblyline_client==v3.7.3


## The following requirements were added by pip freeze:
awscli==1.16.221
bleach==3.1.0
blinker==1.4
botocore==1.12.211
certifi==2019.6.16
chardet==3.0.4
Click==7.0
colorama==0.3.9
dnspython==1.16.0
docutils==0.15.2
flask-redis==0.4.0
greenlet==0.4.15
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
MarkupSafe==1.1.1
mistune==0.8.4
monotonic==1.5
orderedset==2.0.1
phonenumbers==8.10.13
prometheus-client==0.2.0
pyasn1==0.4.6
PyPDF2==1.26.0
python-dateutil==2.8.0
python-json-logger==0.1.11
pytz==2019.1
PyYAML==4.2b1
redis==3.3.8
rsa==3.4.2
s3transfer==0.1.13
six==1.12.0
smartypants==2.0.1
socketIO-client==0.5.6
statsd==3.3.0
urllib3==1.25.3
webencodings==0.5.1
websocket-client==0.56.0
Werkzeug==0.15.5
